pyproject.toml
[project]
name = "ares-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "azure-ai-documentintelligence>=1.0.2",
    "azure-cognitiveservices-speech>=1.45.0",
    "azure-search-documents>=11.5.3",
    "azure-storage-blob>=12.26.0",
    "django>=5.2.5",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "gradio>=5.44.1",
    "markdown>=3.8.2",
    "openai>=1.102.0",
    "pandas>=2.3.2",
    "pdfplumber>=0.11.7",
    "pyarrow>=21.0.0",
    "pypdf>=6.0.0",
    "python-docx>=1.2.0",
    "requests>=2.32.5",
    "tqdm>=4.67.1",
]

[tool.ruff]
line-length = 120
target-version = "py311"
select = [
    "ALL", # include all the rules, including new ones
]
ignore = [
    #### modules
    "ANN", # flake8-annotations
    "COM", # flake8-commas
    "C90", # mccabe complexity
    "DJ",  # django
    "EXE", # flake8-executable
    "T10", # debugger
    "TID", # flake8-tidy-imports

    #### specific rules
    "I001",   # ignore unused imports
    "D100",   # ignore missing docs
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
    "D200",
    "D205",
    "D212",
    "D400",
    "D401",
    "D415",
    "E402",   # false positives for local imports
    "E501",   # line too long
    "TRY003", # external messages in exceptions are too verbose
    "TD002",
    "TD003",
    "FIX002", # too verbose descriptions of todos
]
